♻️ refactor: search 폴더를 search-result로 변경
