+++ components/search-result/map/toggleViewButtonWrapper/index.tsx
