+++ types/search-result/map/type.ts
