+++ types/search-result/toggleViewButton/types.ts
